+++ data/repositories/search/src/main/java/com/sanaa/search/response/SearchResponse.kt
